requirements.txt
pymongo
motor
dnspython
tgcrypto
asyncio
python-dotenv
aiohttp
git+https://github.com/Mayuri-Chan/pyrofork